Everywhere: Move to shared `ValidationException` from scalatra package
