+++ article-api/src/test/scala/no/ndla/articleapi/validation/ContentValidatorTest.scala
@@ -11,7 +11,7 @@ package no.ndla.articleapi.validation
 import no.ndla.articleapi.model.domain._
 import no.ndla.articleapi.{TestEnvironment, UnitSuite}
 import no.ndla.mapping.License.{CC_BY_SA, NA}
-import no.ndla.validation.{ValidationException, ValidationMessage}
+import no.ndla.scalatra.error.{ValidationException, ValidationMessage}
 
 import scala.util.Failure
