+++ draft-api/src/main/scala/draftapi/db/migration/V31__ConvertBrightcoveIds.scala
@@ -10,7 +10,6 @@ package draftapi.db.migration
 import com.typesafe.scalalogging.LazyLogging
 import no.ndla.common.Environment.prop
 import no.ndla.network.model.HttpRequestException
-import no.ndla.validation.{ValidationException, ValidationMessage}
 import org.flywaydb.core.api.migration.{BaseJavaMigration, Context}
 import org.json4s
 import org.json4s.JsonAST.{JArray, JString}
@@ -30,6 +29,7 @@ import scala.collection.mutable.ListBuffer
 import scala.concurrent.duration.Duration
 import scala.concurrent.{Await, ExecutionContext, Future}
 import scala.util.{Failure, Success, Try}
+import no.ndla.scalatra.error.ValidationException
 
 case class BrightcoveToken(access_token: String, expires_in: Long)
 case class StoredToken(accessToken: String, expiresAt: Long)
@@ -88,7 +88,7 @@ class BrightcoveApiClient {
 
   private def extract[T](json: String)(implicit mf: scala.reflect.Manifest[T]): Try[T] = {
     Try { read[T](json) } match {
-      case Failure(e)    => Failure(new ValidationException(errors = Seq(ValidationMessage("body", e.getMessage))))
+      case Failure(e)    => Failure(ValidationException("body", e.getMessage))
       case Success(data) => Success(data)
     }
   }
