+++ draft-api/src/test/scala/no/ndla/draftapi/service/WriteServiceTest.scala
@@ -14,7 +14,7 @@ import no.ndla.draftapi.model.domain.ArticleStatus.{DRAFT, PUBLISHED}
 import no.ndla.draftapi.model.domain._
 import no.ndla.draftapi.model.{api, domain}
 import no.ndla.draftapi.{TestData, TestEnvironment, UnitSuite}
-import no.ndla.validation.{HtmlTagRules, ValidationMessage}
+import no.ndla.validation.{HtmlTagRules}
 import org.mockito.ArgumentMatchers._
 import org.mockito.invocation.InvocationOnMock
 import org.mockito.{ArgumentCaptor, Mockito}
@@ -25,6 +25,7 @@ import java.io.ByteArrayInputStream
 import java.time.LocalDateTime
 import java.util.UUID
 import scala.util.{Failure, Success, Try}
+import no.ndla.scalatra.error.ValidationMessage
 
 class WriteServiceTest extends UnitSuite with TestEnvironment {
   override val converterService = new ConverterService
