+++ validation/src/main/scala/no/ndla/validation/Exceptions.scala
@@ -8,6 +8,3 @@
 package no.ndla.validation
 
 class ConfigurationException(message: String) extends RuntimeException(message)
-
-class ValidationException(message: String = "Validation Error", val errors: Seq[ValidationMessage])
-    extends RuntimeException(message)
