+++ validation/src/main/scala/no/ndla/validation/TagValidator.scala
@@ -14,6 +14,7 @@ import org.jsoup.nodes.{Element, Node}
 
 import scala.jdk.CollectionConverters._
 import scala.util.{Success, Try}
+import no.ndla.scalatra.error.ValidationMessage
 
 class TagValidator {
