+++ validation/src/main/scala/no/ndla/validation/TextValidator.scala
@@ -12,6 +12,7 @@ import org.jsoup.Jsoup
 import org.jsoup.safety.Whitelist
 
 import scala.jdk.CollectionConverters.IteratorHasAsScala
+import no.ndla.scalatra.error.ValidationMessage
 
 class TextValidator(allowHtml: Boolean) {
   private def IllegalContentInBasicText =
