--- validation/src/main/scala/no/ndla/validation/ValidationMessage.scala
@@ -1,10 +1,0 @@
-/*
- * Part of NDLA validation.
- * Copyright (C) 2017 NDLA
- *
- * See LICENSE
- */
-
-package no.ndla.validation
-
-case class ValidationMessage(field: String, message: String)
