+++ validation/src/test/scala/no/ndla/validation/EmbedTagValidatorTest.scala
@@ -10,6 +10,7 @@ package no.ndla.validation
 import no.ndla.mapping.UnitSuite
 import no.ndla.validation.EmbedTagRules.ResourceHtmlEmbedTag
 import no.ndla.validation.TagRules.Condition
+import no.ndla.scalatra.error.ValidationMessage
 
 class EmbedTagValidatorTest extends UnitSuite {
   val embedTagValidator = new TagValidator()
